homework
homework--
my understanding of the assignment is to make an app where the user will be able to create his own quiz, and then reply to the questions. So this the first part to create it.

I have had some problems regarding how to addEventListener to the save options buttons. If you have some tips it'll be greatly appreciated :))!